Updated InternalList property.
I have a collection that inherits from this class, and while debugging, I discovered this property actually returns the opposite of what its current documentation says (at least on .NET 4.6.1, but from looking at the reference source for 4.7.2, it looks like it's still the same). It seems to me like the documentation was just paraphrased from what the base's SourceCollection property says.