Update FileStream example (#1297)
The example given in documentation for invoking a member inconsistent with the purpose of an object actually seems to be very consistent: I would expect to call `FileStream.Read` on a read-only FileStream object and NOT have a `NotSupportedException` thrown. However, if I called `FileStream.Write` on a read-only `FileStream` object, I would expect an exception to be thrown then.