M33065: Possible typo "have the same"/"have not the same" (#1476)
* M33065: Possible typo "have the same"/"have not the same"

Hello, @mairaw ,
Translator has reported possible source content issue. 
Description: Would you please check if the two parts connected with "otherwise" in between are okay as they are; to me, it seems like they are in a meaning clash, as they basically mean the same thing yet "false" and "true."
In other words, I believe - since "{1}" is replaced with "false," - "have the same" should be changed to "do not have the same."

Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.

* Reverted change; made changes to both other overloads.