Mention option of passing null as message parameter for FileNotFoundException
From the current documentation it is not easy to discover that it is possible to get a system-supplied error message containing a provided file name by passing null to the message parameter of the FileNotFoundException constructor. It is mentioned in the Message property, but I only found that when creating this pull request.