The documentation is wrong for Timer.Change method (#3762)
* The documentation is wrong for Timer.Change method

The documentation for the following overload of Timer.Change is incorrect:

public bool Change (long dueTime, long period);

If you use a value for the dueTime parameter that is greater than 4294967294, the method throws an ArgumentOutOfRangeException, not a NotSupportedException as the documentation claims.  I'm not sure which exception is thrown if the period parameter is greater than 4294967294.  I didn't test that.  You might want to check and verify.  You should make it clear in the parameter description that values over 4294967294 are not valid.  We ran into this problem and it took a little while to debug because the documentation was not clear and plain wrong.  Thank you!

* Update xml/System.Threading/Timer.xml

Co-Authored-By: Maira Wenzel <mairaw@microsoft.com>

* Update xml/System.Threading/Timer.xml

Co-Authored-By: Maira Wenzel <mairaw@microsoft.com>

Co-authored-by: Maira Wenzel <mairaw@microsoft.com>