Corrected documentation for overrides with an overwriteFiles parameter (#6675)
Changed IOException to state that it is only thrown if overwriteFiles is false and the specified directory does not exist.
Changed the remarks to reflect that change.  
Removed statement in the description for the destinationDirectoryName param that said directory must exist.
Specified in remarks that both sourceArchiveFileName and destinationDirectoryName param accept relative and absolute paths.